+++ src/avp/bh_snds.h
@@ -17,11 +17,11 @@ typedef struct sound_tools_template
 {
 	VECTORCH position;
 
-	unsigned long inner_range;
-	unsigned long outer_range;
+    uint32_t inner_range;
+    uint32_t outer_range;
 	
-	unsigned long max_volume;
-	unsigned long pitch;
+    uint32_t max_volume;
+    uint32_t pitch;
 	
 	unsigned int playing :1;
 	unsigned int loop :1;
@@ -36,8 +36,8 @@ typedef struct sound_behav_block
 {
 	VECTORCH position;
 	
-	unsigned long inner_range;
-	unsigned long outer_range;
+    uint32_t inner_range;
+    uint32_t outer_range;
 	int max_volume;
 	int	pitch;
